requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
# psycopg2-binary==2.9.9
google-auth==2.23.4
boto3==1.29.7
python-decouple==3.8
django-cors-headers==4.3.1
phonenumbers==8.13.26